encapsulations
