+++ src/main/java/ortus/boxlang/runtime/cli/MiniConsole.java
@@ -141,81 +141,6 @@ public class MiniConsole implements AutoCloseable {
 	}
 
 	/**
-	 * Internal class to track completion state for cycling through suggestions
-	 */
-	private static class CompletionState {
-
-		private final String				originalInput;
-		private final List<TabCompletion>	completions;
-		private final int					cursorPosition;
-		private int							currentIndex;
-		private final int					replaceStart;
-		private final int					replaceEnd;
-
-		CompletionState( String originalInput, List<TabCompletion> completions, int cursorPosition, int replaceStart, int replaceEnd ) {
-			this.originalInput	= originalInput;
-			this.completions	= new ArrayList<>( completions );
-			this.cursorPosition	= cursorPosition;
-			this.currentIndex	= 0;
-			this.replaceStart	= replaceStart;
-			this.replaceEnd		= replaceEnd;
-		}
-
-		TabCompletion getCurrentCompletion() {
-			return this.completions.isEmpty() ? null : this.completions.get( currentIndex );
-		}
-
-		void nextCompletion() {
-			if ( !this.completions.isEmpty() ) {
-				currentIndex = ( currentIndex + 1 ) % this.completions.size();
-			}
-		}
-
-		boolean hasCompletions() {
-			return !this.completions.isEmpty();
-		}
-
-		int getCompletionCount() {
-			return this.completions.size();
-		}
-
-		/**
-		 * Get all completions in this state
-		 */
-		List<TabCompletion> getAllCompletions() {
-			return this.completions;
-		}
-
-		/**
-		 * Get the current completion index
-		 */
-		int getCurrentIndex() {
-			return this.currentIndex;
-		}
-
-		/**
-		 * Get the original input before any completions
-		 */
-		String getOriginalInput() {
-			return this.originalInput;
-		}
-
-		/**
-		 * Get the replace start position
-		 */
-		int getReplaceStart() {
-			return this.replaceStart;
-		}
-
-		/**
-		 * Get the replace end position
-		 */
-		int getReplaceEnd() {
-			return this.replaceEnd;
-		}
-	}
-
-	/**
 	 * Constructor with custom prompt
 	 *
 	 * @param prompt The prompt string to display
